refactor: split rest react helpers
